Update confirmed detail mock
