+++ src/applications/vaos/containers/ConfirmedAppointmentPage.jsx
@@ -2,26 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 import { connect } from 'react-redux';
-import moment from 'moment';
 import LoadingIndicator from '@department-of-veterans-affairs/formation-react/LoadingIndicator';
 import { focusElement } from 'platform/utilities/ui';
 import { fetchConfirmedAppointments } from '../actions/appointments';
-import { FETCH_STATUS, PURPOSE_TEXT } from '../utils/constants';
+import { FETCH_STATUS } from '../utils/constants';
 import { selectConfirmedAppointment } from '../utils/selectors';
-import { formatTimeToCall } from '../utils/formatters';
-
-function formatDate(date) {
-  const parsedDate = moment(date, 'MM/DD/YYYY hh:mm:ss');
-
-  if (!parsedDate.isValid()) {
-    return '';
-  }
-
-  return {
-    date: parsedDate.format('MMMM D, YYYY'),
-    time: parsedDate.format('h:mm a'),
-  };
-}
+import { getTitle, getLocation, getDateTime } from '../utils/appointment';
 
 export class ConfirmedAppointmentPage extends React.Component {
   componentDidMount() {
@@ -30,10 +16,6 @@ export class ConfirmedAppointmentPage extends React.Component {
   }
   render() {
     const { appointment, status } = this.props;
-
-    const formattedDateTime = appointment?.bookedApptDateTime
-      ? formatDate(appointment.bookedApptDateTime)
-      : {};
 
     return (
       <div className="vads-l-grid-container vads-u-padding-x--2p5 large-screen:vads-u-padding-x--0 vads-u-padding-bottom--2p5">
@@ -50,36 +32,17 @@ export class ConfirmedAppointmentPage extends React.Component {
             )}
             {status === FETCH_STATUS.succeeded && (
               <>
-                <h2>{appointment.appointmentType}</h2>
+                <h2>{getTitle(appointment)}</h2>
                 <div className="vads-u-display--flex vads-u-margin-bottom--2">
                   <div className="vads-u-flex--1">
                     <>
                       <h3 className="vaos-appts__block-label">Where</h3>
-                      {appointment.friendlyLocationName ||
-                        appointment.facility.name}
-                      <br />
-                      {appointment.facility.city}, {appointment.facility.state}
+                      {getLocation(appointment)}
                     </>
                     <h3 className="vaos-appts__block-label vads-u-margin-top--2">
                       When
                     </h3>
-                    {formattedDateTime.date}
-                    <br />
-                    {formattedDateTime.time}
-                    <h3 className="vaos-appts__block-label vads-u-margin-top--2">
-                      Purpose
-                    </h3>
-                    {PURPOSE_TEXT[appointment.purposeOfVisit] ||
-                      appointment.reasonForVisit}
-                    <h3 className="vaos-appts__block-label vads-u-margin-top--2">
-                      Type
-                    </h3>
-                    {appointment.visitType}
-                    <h3 className="vaos-appts__block-label vads-u-margin-top--2">
-                      My contact number
-                    </h3>
-                    {appointment.phoneNumber}, in the{' '}
-                    {formatTimeToCall(appointment.bestTimetoCall)}
+                    {getDateTime(appointment)}
                   </div>
                 </div>
                 <Link to="appointments/confirmed">
