+++ src/applications/vaos/utils/selectors.js
@@ -1,6 +1,10 @@
+import { getAppointmentId } from './appointment';
+
 export function selectConfirmedAppointment(state, id) {
   return (
-    state.appointments?.confirmed?.find?.(appt => appt.uniqueId === id) || null
+    state.appointments?.confirmed?.find?.(
+      appt => getAppointmentId(appt) === id,
+    ) || null
   );
 }
